refactor: Move functionFinding-functions into Polymorphism module
