+++ CarpHask.cabal
@@ -29,7 +29,6 @@ library
                        EvalError,
                        Eval,
                        Expand,
-                       FindFunction,
                        Forms,
                        GenerateConstraints,
                        Infer,
