+++ src/ArrayTemplates.hs
@@ -3,8 +3,8 @@
 module ArrayTemplates where
 
 import Concretize
-import FindFunction
 import Obj
+import Polymorphism
 import StructUtils
 import Template
 import ToTemplate
